+++ scripts/interface_llm.py
@@ -131,12 +131,13 @@ def terminate_server(verbosity):
     global PROCESS
 
     if PROCESS is not None:
-        if verbosity > 1:
-            print('Terminating llm server (may freeze OS for a few seconds during mem free)')
+        if PROCESS.poll() is None:
+            if verbosity > 1:
+                print('Terminating llm server (may freeze OS for a few seconds during mem free)')
 
-        # cannot use Popen.terminate() or kill() because they will not kill further spawned processes, especially the process responsible for consuming vram
-        os.killpg(os.getpgid(PROCESS.pid), signal.SIGTERM)
-        PROCESS.communicate(timeout=60)  # clears pipe buffers, and waits for process to close
+            # cannot use Popen.terminate() or kill() because they will not kill further spawned processes, especially the process responsible for consuming vram
+            os.killpg(os.getpgid(PROCESS.pid), signal.SIGTERM)
+            PROCESS.communicate(timeout=60)  # clears pipe buffers, and waits for process to close
 
         PROCESS = None
